Fix definitions and add readme
