+++ src/config.py
@@ -1,0 +1,1 @@
+INDENT: str = "\t"
